Reorganize contrib.epidemiology models (#2499)
